feat: declarate search()
This function is called from the show_output(event) and receive one argument: word. Then open the archive python.txt, who contain different concepts. The function tries to search the word in the archive. If the word exists, returns a list with the information's about the search, if not, return a "No" string.