feat: declared def show_output(event)
This function is called after the input into our search concept inside the Entry widget and activated by pressing return. Here, we create a variable called word who receive the input and is passed as an argument to the search() function. If we receive a "No" in return, then we obtain a fail message. In case of a yes, it returns text answers one to seven.